+++ tests/dmd/runnable/test17559.d
@@ -1,7 +1,5 @@
 // REQUIRED_ARGS: -g
 // REQUIRED_ARGS(linux freebsd openbsd dragonflybsd): -L-export-dynamic
-// LDC (required for Win32 and -O): REQUIRED_ARGS(windows32): -link-defaultlib-debug
-// LDC (FreeBSD's libexecinfo apparently doesn't like elided frame pointers): REQUIRED_ARGS(freebsd): -link-defaultlib-debug -frame-pointer=all
 // PERMUTE_ARGS:
 // DISABLED: osx